chore(boot): Moving system utils into boot
