+++ .envrc
@@ -3,10 +3,10 @@ while [[ ! -e ".git" ]]; do
   ROOT=`dirname $ROOT`
 done
 export PATH="$PATH:$ROOT/.bin"
-export NODE_OPTIONS="-r $ROOT/module/boot/src/internal/dev/register.js"
 export NODE_PATH="$NODE_PATH:$ROOT/.bin"
 
 export TRV_DOC_BRANCH=master
 export TRV_DEV_ROOT=$ROOT
 export TRV_DEV=$ROOT/module
+export TRV_REQUIRES="$ROOT/.bin/dev-register.js"
 export TRV_CLI_VERIFY=0
