+++ module/boot/bin/register.js
@@ -7,4 +7,5 @@ if (global[s]) {
 }
 global[s] = __filename;
 try { require(`${process.cwd()}/.env`); } catch { } // read env
+(process.env.TRV_REQUIRES || '').replace(/[^ ,:]+/g, f => require(f));
 require('@travetto/boot/src/internal/module').ModuleManager.init(); // init
