+++ module/boot/src-ts/internal/system.ts
@@ -1,0 +1,18 @@
+/**
+ * Shared util
+ */
+export class SystemUtil {
+  /**
+   * Naive hashing
+   */
+  static naiveHash(text: string) {
+    let hash = 5381;
+
+    for (let i = 0; i < text.length; i++) {
+      // eslint-disable-next-line no-bitwise
+      hash = (hash * 33) ^ text.charCodeAt(i);
+    }
+
+    return Math.abs(hash);
+  }
+}
