+++ module/doc/doc.ts
@@ -1,0 +1,6 @@
+import { Header } from './src/nodes';
+import { doc as d } from './src/helper';
+
+export const text = d`
+${Header()}
+`;
