+++ module/log/src/service.ts
@@ -1,6 +1,6 @@
 import { EnvUtil } from '@travetto/boot';
 import { ConsoleManager, LogLevel, AppManifest, Util } from '@travetto/base';
-import { SystemUtil } from '@travetto/base/src/internal/system';
+import { ModuleUtil } from '@travetto/boot/src/internal/module-util';
 
 import { Appender, Formatter, LogEvent, LogLevels } from './types';
 import { LineFormatter } from './formatter/line';
@@ -123,7 +123,7 @@ class $Logger {
       message = '';
     }
 
-    const category = SystemUtil.computeModule(file);
+    const category = ModuleUtil.getId(file);
 
     if ((level in this.exclude) || (category && level in this.filters && !this.filters[level]!(category))) {
       return;
