+++ module/model-elasticsearch/test/schema.ts
@@ -101,7 +101,6 @@ class SchemaSuite {
 
   @Test('Numeric schema')
   async testNumericSchema() {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const schema3: any = ElasticsearchSchemaUtil.generateSourceSchema(Numerical);
 
     assert(schema3.properties.money.type === 'scaled_float');
