+++ module/model-query/test/parser.ts
@@ -28,8 +28,7 @@ export class QueryStringTest {
       { type: 'operator', value: '==' },
       { type: 'identifier', value: 'D' },
     ]);
-    // eslint-disable-next-line @typescript-eslint/quotes
-    assert(QueryLanguageTokenizer.tokenize(`"A.b.c"   =='  D'`) === [
+    assert(QueryLanguageTokenizer.tokenize(`"${'A.b.c'}"   =='  D'`) === [
       { type: 'literal', value: 'A.b.c' },
       { type: 'operator', value: '==' },
       { type: 'literal', value: '  D' }
@@ -172,13 +171,11 @@ export class QueryStringTest {
 
   @Test('Parse Regex')
   async parseRegex() {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const res: any = QueryLanguageParser.parseToQuery<UserType>('user.role ~ /^admin/');
     assert(res === { user: { role: { $regex: /^admin/ } } });
     assert(res.user.role.$regex instanceof RegExp);
     assert(res.user.role.$regex.toString() === '/^admin/');
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const res2: any = QueryLanguageParser.parseToQuery<UserType>("user.role ~ 'admin'");
     assert(res2 === { user: { role: { $regex: /^admin/ } } });
     assert(res2.user.role.$regex instanceof RegExp);
@@ -187,7 +184,6 @@ export class QueryStringTest {
 
   @Test('Parse Regex with flags')
   async parseRegexWithFlags() {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const res: any = QueryLanguageParser.parseToQuery<UserType>('user.role ~ /^admin/i');
     assert(res === { user: { role: { $regex: /^admin/i } } });
     assert(res.user.role.$regex instanceof RegExp);
@@ -200,7 +196,6 @@ export class QueryStringTest {
 
   @Test('Parse Regex with word boundaries')
   async parseRegexWithWordBoundaries() {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const res: any = QueryLanguageParser.parseToQuery<UserType>('user.role ~ /\badmin\b/i');
     assert(res.user.role.$regex instanceof RegExp);
     assert(res.user.role.$regex.toString() === '/\badmin\b/i');
